health(criteria): #80 some file renames & fn to create token
