+++ src/criteria/parser/index.ts
@@ -1,7 +1,9 @@
 export * from "./lex.fn";
+export * from "./parse-criteria.fn";
 export * from "./scan-number.fn";
 export * from "./scan-string.fn";
 export * from "./scan-symbol.fn";
+export * from "./token-parser";
 export * from "./token-type.enum";
-export * from "./token";
-export * from "./parse-criteria.fn";
+export * from "./token.contract";
+export * from "./token.fn";
