+++ src/criteria/parser/token-parser/index.ts
@@ -1,5 +1,5 @@
-export * from "./parse-in-range-generator.fn";
-export * from "./parse-in-set-generator.fn";
+export * from "./parse-in-range.generator";
+export * from "./parse-in-set.generator";
 export * from "./parse-token-generator.type";
-export * from "./parse-value-criteria-generator.fn";
-export * from "./parse-value-criterion-generator.fn";
+export * from "./parse-value-criteria.generator";
+export * from "./parse-value-criterion.generator";
