+++ src/criteria/parser/token-parser/parse-token-generator.type.ts
@@ -1,4 +1,4 @@
 import { ValueCriterion } from "../../value-criterion";
-import { Token } from "../token";
+import { Token } from "../token.contract";
 
 export type ParseTokenGenerator = Generator<boolean, false | ValueCriterion, Token>;
