+++ src/criteria/parser/token.contract.ts
