+++ src/criteria/parser/token.fn.ts
@@ -1,0 +1,6 @@
+import { TokenType } from "./token-type.enum";
+import { Token } from "./token.contract";
+
+export function token(type: TokenType, value: string): Token {
+    return { type, value };
+}
