+++ src/criteria/test/parsing/criteria-lexer.spec.ts
@@ -1,4 +1,4 @@
-import { lex, Token, TokenType } from "../../parser";
+import { lex, token, Token, TokenType } from "../../parser";
 
 describe("criteria-lexer", () => {
     it("should return empty array of tokens when given an empty string", () => {
@@ -26,18 +26,14 @@ describe("criteria-lexer", () => {
         author: { name: {"bob"} } | { surname: {"uncle", "burger"} }
     }`;
 
-    shouldLex("123", [{ type: TokenType.Number, value: "123" }]);
-    shouldLex("{123}", [
-        { type: TokenType.Special, value: "{" },
-        { type: TokenType.Number, value: "123" },
-        { type: TokenType.Special, value: "}" },
-    ]);
+    shouldLex("123", [token(TokenType.Number, "123")]);
+    shouldLex("{123}", [token(TokenType.Special, "{"), token(TokenType.Number, "123"), token(TokenType.Special, "}")]);
 
     shouldLex("[123, 789)", [
-        { type: TokenType.Special, value: "[" },
-        { type: TokenType.Number, value: "123" },
-        { type: TokenType.Special, value: "," },
-        { type: TokenType.Number, value: "789" },
-        { type: TokenType.Special, value: ")" },
+        token(TokenType.Special, "["),
+        token(TokenType.Number, "123"),
+        token(TokenType.Special, ","),
+        token(TokenType.Number, "789"),
+        token(TokenType.Special, ")"),
     ]);
 });
